requirements.txt
Pillow
PyYAML
opencv-python
numpy
python-docx